Has solve method to solve CuttingStock object
Refactors the cuttingStockProblem function in the cutting_stock.py script to be a method for the CuttingStock class. Ideally this would be in the class script, but it's pretty long so I made it a separate script and import it in the class. Note that the method has self as a parameter.